`Animation` -> `KeyFrames`
Decided instead to go with just a `KeyFrames` type with base `ToolipsServables`, instead opting for the transition animations to be done solely with a `ComponentModifier`, or other `Modifier`. This is getting really close. I just need to finish the last few aspects of `StyleComponents` With this, I will wrap up the tests and slowly trickle into a grand `Toolips` release.  I plan to do this pretty soon, I have been working on a game (emperium) https://github.com/emmaccode/emperium) and this has absorbed a lot of the free time from other projects, but I plan to migrate the focus back here a bit